move annotator base to annotators directory
