+++ ansible_risk_insight/annotators/ansible_builtin.py
@@ -17,8 +17,8 @@
 import json
 from typing import List
 from ..models import TaskCall, Annotation, RiskAnnotation, VariableAnnotation
-from ..variable_resolver import VARIABLE_ANNOTATION_TYPE
-from .base import RiskAnnotator, RiskType
+from .variable_resolver import VARIABLE_ANNOTATION_TYPE
+from .risk_annotator_base import RiskAnnotator, RiskType
 
 
 class AnsibleBuiltinRiskAnnotator(RiskAnnotator):
